+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 1.1.0 - 2026-01-22
+
 ### Added
 
 - Modem GPS drawer toggle